Add registration functionality
The update primarily aims to incorporate a registration feature to the application. In the changes, the ability for a user to input their username, email address, password and password confirmation through text fields has been added. Furthermore, the user's input is processed by the RegisterBloc, with individual events triggering for each input field. The RegisterController handles the backend interaction, specifically invoking the FirebaseAuth API to create a user account based on the entered data. Errors in the forms of weak passwords, account existence and invalid emails are detailed through error messages to ensure efficient feedback. RouteNames for this new feature has also been properly set up. Lastly, the file ApplicationPage has been created to serve as a main router. Purpose of this registration feature is to enable user-specific customizations, histories and experiences.